Updated metadata.rb, and fcontext to allow rake to run.
Wrote a lot of tests in port_spec.rb to demonstrate the issue, and to prove what I have now works.

The problem was using selinux_policy_port on a port that is defined in a system range, for example amanda_port_t in the 10080-10083 range.  When attempting to add here, it would try to modify instead of add because it detected the range of ports.
Now it will correctly detect if there is a range of ports in a different context covering the port you're adding, and force the add instead of modify.

It appears the issue was really with attempting to add to ports already defined in a range.  If you want to add to a range, you have to add not modify.

Now it will detect if a range of ports exists, and force these to be added when one comes along of a different context instead of modify.

Signed-off-by: Bryan Killian <bryan.killian@ardentmc.com>